git-svn-id: https://svn.code.sf.net/p/kaldi/code/sandbox/dan@1611 5e6a8d80-dfce-4ca6-a32a-6e07a63d50c8

+++ src/nnet/nnet-cache.h
@@ -73,11 +73,11 @@ class Cache {
   
   State state_; ///< Current state of the cache
 
-  size_t filling_pos_;  ///< Number of frames filled to cache by AddData
-  size_t emptying_pos_; ///< Number of frames given by cache by GetBunch
+  int32 filling_pos_;  ///< Number of frames filled to cache by AddData
+  int32 emptying_pos_; ///< Number of frames given by cache by GetBunch
   
-  size_t cachesize_; ///< Size of cache
-  size_t bunchsize_; ///< Size of bunch
+  int32 cachesize_; ///< Size of cache
+  int32 bunchsize_; ///< Size of bunch
 
   bool randomized_;
